Move AsyncCounter into separate file
